Fix build, move stuff to private, use str
