+++ src/private/cstr.rs
@@ -6,6 +6,8 @@ pub use alloc::ffi::CString;
 
 pub use core::ffi::{c_char, CStr};
 
+use crate::sys::{EspError, ESP_ERR_INVALID_SIZE};
+
 #[cfg(feature = "alloc")]
 pub fn set_str(buf: &mut [u8], s: &str) -> Result<(), crate::sys::EspError> {
     assert!(s.len() < buf.len());
@@ -77,4 +79,110 @@ pub fn nul_to_invalid_arg(_err: alloc::ffi::NulError) -> crate::sys::EspError {
 #[cfg(feature = "alloc")]
 pub fn to_cstring_arg(value: &str) -> Result<CString, crate::sys::EspError> {
     CString::new(value).map_err(nul_to_invalid_arg)
+}
+
+/// str to cstr, will be truncated if str is larger than buf.len() - 1
+///
+/// # Panics
+///
+/// * Panics if buffer is empty.
+pub fn cstr_from_str_truncating<'a>(rust_str: &str, buf: &'a mut [u8]) -> &'a CStr {
+    assert!(buf.len() > 0);
+
+    let max_str_size = buf.len() - 1; // account for NUL
+    let truncated_str = &rust_str[..max_str_size.min(rust_str.len())];
+    buf[..truncated_str.len()].copy_from_slice(truncated_str.as_bytes());
+    buf[truncated_str.len()] = b'\0';
+
+    CStr::from_bytes_with_nul(&buf[..truncated_str.len() + 1]).unwrap()
+}
+
+/// Convert slice of rust strs to NULL-terminated fixed size array of c string pointers
+///
+/// # Panics
+///
+/// * Panics if cbuf is empty.
+/// * Panics if N is <= 1
+pub fn cstr_arr_from_str_slice<const N: usize>(
+    rust_strs: &[&str],
+    mut cbuf: &mut [u8],
+) -> Result<[*const i8; N], EspError> {
+    assert!(N > 1);
+    assert!(cbuf.len() > 0);
+
+    // ensure last element stays NULL
+    if rust_strs.len() > N - 1 {
+        return Err(EspError::from_infallible::<ESP_ERR_INVALID_SIZE>());
+    }
+
+    let mut cstrs = [std::ptr::null(); N];
+
+    for (i, s) in rust_strs.into_iter().enumerate() {
+        let max_str_size = cbuf.len() - 1; // account for NUL
+        if s.len() > max_str_size {
+            return Err(EspError::from_infallible::<ESP_ERR_INVALID_SIZE>());
+        }
+        cbuf[..s.len()].copy_from_slice(s.as_bytes());
+        cbuf[s.len()] = b'\0';
+        let cstr = CStr::from_bytes_with_nul(&cbuf[..s.len() + 1]).unwrap();
+        cstrs[i] = cstr.as_ptr();
+
+        cbuf = &mut cbuf[s.len() + 1..];
+    }
+
+    Ok(cstrs)
+}
+
+#[cfg(test)]
+mod tests {
+    use std::ffi::CStr;
+
+    use crate::{cstr_arr_from_str_slice, cstr_from_str};
+
+    #[test]
+    fn cstr_from_str_happy() {
+        let mut same_size = [0u8; 6];
+        let hello = cstr_from_str("Hello", &mut same_size);
+        assert_eq!(hello.to_bytes(), b"Hello");
+
+        let mut larger = [0u8; 42];
+        let hello = cstr_from_str("Hello", &mut larger);
+        assert_eq!(hello.to_bytes(), b"Hello");
+    }
+
+    #[test]
+    fn cstr_from_str_unhappy() {
+        let mut smaller = [0u8; 6];
+        let hello = cstr_from_str("Hello World", &mut smaller);
+        assert_eq!(hello.to_bytes(), b"Hello");
+    }
+
+    #[test]
+    fn cstr_arr_happy() {
+        let mut same_size = [0u8; 13];
+        let hello = cstr_arr_from_str_slice::<3>(&["Hello", "World"], &mut same_size).unwrap();
+        assert_eq!(unsafe { CStr::from_ptr(hello[0]) }.to_bytes(), b"Hello");
+        assert_eq!(unsafe { CStr::from_ptr(hello[1]) }.to_bytes(), b"World");
+        assert_eq!(hello[2], std::ptr::null());
+    }
+
+    #[test]
+    #[should_panic]
+    fn cstr_arr_unhappy_n1() {
+        let mut cbuf = [0u8; 25];
+        let _ = cstr_arr_from_str_slice::<1>(&["Hello"], &mut cbuf);
+    }
+
+    #[test]
+    fn cstr_arr_unhappy_n_too_small() {
+        let mut cbuf = [0u8; 25];
+        assert!(cstr_arr_from_str_slice::<2>(&["Hello", "World"], &mut cbuf).is_err());
+    }
+
+    #[test]
+    #[should_panic]
+    fn cstr_arr_unhappy_cbuf_too_small() {
+        let mut cbuf = [0u8; 12];
+        assert!(cstr_arr_from_str_slice::<3>(&["Hello", "World"], &mut cbuf).is_err());
+    }
 }
